requirements.txt
colorama==0.4.1
ruamel.yaml==0.15.83

psycopg2==2.7.6.1

requests==2.21.0
beautifulsoup4==4.6.3
lxml==4.3.0

numpy==1.15.4
pandas==0.23.4
